dom/tests/mochitest/mouselock/test_defaultUnlock.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Test for Bug 633602</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
	<script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
	<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>

	<p id="display"></p>

	<div id="content" style="display: none">
	</div>
	
	<canvas id="canvas" width="150" height="150"></canvas>
	<pre id="test">
    <script>
		  SimpleTest.waitForExplicitFinish();
      SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only", false);

      const canvas = document.getElementById("canvas");
      const pointer = navigator.pointer;
      var runOnce = false;

      var onLockGet = function (e) {
        ok(pointer.islocked(), "Element should be locked after the pointer's lock() function is called");
        synthesizeKey("VK_ESCAPE", { type: "keydown" }, window);
        synthesizeKey("VK_ESCAPE", { type: "keyup" }, window);
      };
      
      var onLockFail = function (e) {
        ok(false, "Lock()'s failed callback was called, lock() did not succeed");
        SimpleTest.finish();
      };
      
      var onFullScreen = function (e) {
        if (!runOnce) {
          runOnce = true;
          pointer.lock(canvas, onLockGet, onLockFail);
        } else {
          ok(!pointer.islocked(), "Element should be unlocked after ESC key was injected");
          SimpleTest.finish();
        }
      };

      document.addEventListener("mozfullscreenchange", onFullScreen, false);

      canvas.mozRequestFullScreen();
		</script>
	</pre>
</body>
</html>
